# Improve performance of SQL proxy_configs query