## NEW ⚠️ 
@akostadinov came up with a new query that is also very quick and it can be still a scope, therefore we can keep the `.current_versions` format.

## OLD, don't read 💩 

The query to be fixed is `ProxyConfig#current_versions`. It is a very complex one because it's supposed to check the ProxyConfig with max version for each Proxy and each environment, and be compatible with other scopes such as `by_environment` and `by_host`.

After some investigation and testing, this is **not possible**. The query is just too complex for ActiveRecord to figure out, therefore I've refactored it into a class method that will run a handcrafted query, hopefully maintaining the same behaviour.

Here are the places where `current_versions` is called and a comparison with the old and new queries, using 7000 Proxy configs.

### Case A
This endpoint is not in the doc and it's supposed to be removed at some point (see #2289):
https://github.com/3scale/porta/blob/b9757da19cce1d5ed6c859b9ec6cc3af68d53f2d/app/controllers/admin/api/services/proxy/configs_controller.rb#L112-L115
```sql
-- ProxyConfig Load (753185.4ms)
SELECT proxy_configs.* FROM proxy_configs
INNER JOIN proxies ON proxy_configs.proxy_id = proxies.id
INNER JOIN services ON proxies.service_id = services.id
INNER JOIN proxy_configs versions ON versions.proxy_id = proxy_configs.proxy_id
  AND versions.environment = proxy_configs.environment
WHERE services.account_id = 2
  AND (services.state != 'deleted')
  AND proxy_configs.environment = 'sandbox'
  AND (proxy_config.hosts LIKE '%|pepe.prod.org|%')
GROUP BY proxy_configs.id
HAVING max(versions.version) = proxy_configs.version
```
**New query:**
```ruby
ProxyConfig.current_versions(env: environment, host: host, account: account)
```
```sql
-- ProxyConfig Load (19.4ms)
SELECT * FROM proxy_configs
WHERE (proxy_id, environment, version) IN (
  SELECT proxy_id, environment, MAX(version)
  FROM proxy_configs
  INNER JOIN proxies ON proxies.id = proxy_configs.proxy_id
  INNER JOIN services ON proxies.service_id = services.id
  WHERE services.account_id = "2"
    AND services.state != 'deleted'
    AND proxy_configs.environment = 'sandbox'
  GROUP BY environment, proxy_id
)
AND hosts LIKE '%|pepe.prod.org|%'
```
**Notes**
The new implementation seems to return the same results for this case. The `account_id` and `state != 'deleted'` corresponds to `accessible_services`. In the new query, the host is checked after any maxes are found on purpose: if the host doesn't have the latest version of a service, don't show it.

### Case B
https://github.com/3scale/porta/blob/b9757da19cce1d5ed6c859b9ec6cc3af68d53f2d/app/controllers/master/api/proxy/configs_controller.rb#L21-L23
```sql
-- ProxyConfig Load (764104.5ms)
SELECT proxy_configs.* FROM proxy_configs
INNER JOIN proxy_configs versions ON versions.proxy_id = proxy_configs.proxy_id
  AND versions.environment = proxy_configs.environment
WHERE proxy_configs.environment = 'sandbox'
AND (proxy_config.hosts LIKE '%|pepe.prod.org|%')
GROUP BY proxy_configs.id
HAVING max(versions.version) = proxy_configs.version
```
**New query:**
```ruby
ProxyConfig.current_versions(env: environment, host: host)
```
```sql
-- ProxyConfig Load (24.2ms)
SELECT * FROM proxy_configs
WHERE (proxy_id, environment, version) IN (
  SELECT proxy_id, environment, MAX(version)
  FROM proxy_configs
  WHERE proxy_configs.environment = 'sandbox'
  GROUP BY environment, proxy_id
)
AND hosts LIKE '%|pepe.prod.org|%'
```
**Notes**
...

### Case C
To generate the curl command that's displayed in a Product config page
https://github.com/3scale/porta/blob/b9757da19cce1d5ed6c859b9ec6cc3af68d53f2d/app/lib/apicast/curl_command_builder.rb#L136-L139
```sql
-- ProxyConfig Load (38076.4ms)
SELECT proxy_configs.* FROM proxy_configs
INNER JOIN proxy_configs versions ON versions.proxy_id = proxy_configs.proxy_id
  AND versions.environment = proxy_configs.environment
WHERE proxy_configs.proxy_id = 106
  AND proxy_configs.environment = 'sandbox'
GROUP BY proxy_configs.id
HAVING max(versions.version) = proxy_configs.version
ORDER BY proxy_configs.id ASC
LIMIT 1
```
**New query:**
```ruby
ProxyConfig.current_versions(env: environment, account: proxy.service.account).first
```
```sql
-- ProxyConfig Load (26.4ms)
SELECT * FROM proxy_configs
WHERE (proxy_id, environment, version) IN (
  SELECT proxy_id, environment, MAX(version) FROM proxy_configs
  INNER JOIN proxies ON proxies.id = proxy_configs.proxy_id
  INNER JOIN services ON proxies.service_id = services.id
  WHERE services.account_id = "2"
    AND services.state != 'deleted'
    AND proxy_configs.environment = 'sandbox'
  GROUP BY environment, proxy_id
)
```
**Notes**
The can probably be refactored/simplified instead of using current_versions.


### Case D
Used when filtering by `version = lastest` which is pretty bad, because it's not really compatible with other scopes. `by_version` is used in 1 endpoint only and it's probably better to check for `'later'` manually, outside the scope.
https://github.com/3scale/porta/blob/b9757da19cce1d5ed6c859b9ec6cc3af68d53f2d/app/models/proxy_config.rb#L50-L55
```sql
-- ProxyConfig Load (710401.8ms)
SELECT proxy_configs.* FROM proxy_configs
INNER JOIN proxies ON proxies.id = proxy_configs.proxy_id
INNER JOIN proxy_configs versions ON versions.proxy_id = proxy_configs.proxy_id
  AND versions.environment = proxy_configs.environment
WHERE proxies.service_id IN (2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109)
  AND proxy_configs.environment = 'sandbox'
GROUP BY proxy_configs.id
HAVING max(versions.version) = proxy_configs.version
```
**New query:**
```ruby
ProxyConfig.current_versions(env: environment, host: host, account: account)
```
```sql
-- ProxyConfig Load (21.9ms)
SELECT * FROM proxy_configs
WHERE (proxy_id, environment, version) IN (
  SELECT proxy_id, environment, MAX(version) FROM proxy_configs
  INNER JOIN proxies ON proxies.id = proxy_configs.proxy_id
  INNER JOIN services ON proxies.service_id = services.id
  WHERE services.account_id = "2"
    AND services.state != 'deleted'
    AND proxy_configs.environment = 'sandbox'
  GROUP BY environment, proxy_id
)
```
**Notes**
As far as I can tell this query is the same use case as case A above.

**JIRA:**
[THREESCALE-7183: mysql slow queries](https://issues.redhat.com/browse/THREESCALE-7183)