@raelga 2022-02-25
Just to double-check the service filter can really be skipped safely?

```
`services`.`account_id` = 2 AND (`services`.`state` != 'deleted') ` 
```