@josemigallas 2022-03-08
> Just to double-check the service filter can really be skipped safely?
> 
> ```
> `services`.`account_id` = 2 AND (`services`.`state` != 'deleted') ` 
> ```

Good catch @raelga, I had to add the joins in the end... but the performance of the new query isn't affected as far as I can see.