@slopezz 2022-03-08
@josemigallas what about the `account` verification?

It seems there is no `account` verification in the new query. Could it maybe lead to data leak of customers asking proxy-configs from other customers?