@akostadinov 2022-03-08
One preliminary question - have you tested oracle and postgresql?

Maybe we need to analyze original query for missing table indices.